RAG/requirements_txt.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-cloud-bigquery==3.12.0
PyPDF2==3.0.1
chardet==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
openai==1.3.5
langchain==0.0.350
langchain-community==0.0.1
chromadb==0.4.18
tiktoken==0.5.2
python-dotenv==1.0.0
logging==0.4.9.6